+++ internal/updater/timeouts.go
@@ -13,7 +13,7 @@ func BatchTimeout(cat model.Category) time.Duration {
 	switch cat {
 	case model.CatBrew:
 		return 25 * time.Minute
-	case model.CatMAS, model.CatApt, model.CatPacman:
+	case model.CatMAS, model.CatApt, model.CatPacman, model.CatDnf, model.CatZypper, model.CatApk:
 		return 30 * time.Minute
 	case model.CatWinget, model.CatChoco:
 		return 30 * time.Minute